Clase cartastipo y correccion cartas jugadores
Se ha reestructurado la clase cartastipo para incluirla en un fichero .h y además se ha corregido el error que permitía que algunos jugadores tuviesen las mismas cartas
