+++ cartastipo.h
@@ -1,0 +1,17 @@
+#ifndef CARTASTIPO_H
+#define CARTASTIPO_H
+#include <iostream>
+#include <QString>
+
+using namespace std;
+
+class cartastipo{
+public:
+    cartastipo();
+    int numero;
+    QString tipo;
+    void infodecarta(int a);
+
+};
+
+#endif // CARTASTIPO_H
