+++ poker.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 4.8.1, 2019-04-26T17:30:19. -->
+<!-- Written by QtCreator 4.8.1, 2019-04-29T13:11:27. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
